dynamic: Added 'cmds/dynamic.c'
Parse the command line argument value and inject 'libmcount-dynamic.so'
into the process with the given pid. It also saves environment variables
to a file and makes it available in 'libmcount-dynamic.so'.

cmd-dynamic shares both 'build_debug_domain_string' and
'check_linux_schedule_event' functions with cmd-record. removed
static from the declarations of both functions so that we could
link the functions in cmd-record to reduce duplicate code.

command and option that associated with dynamic trace added to 'uftrace'.

Signed-off-by: Hanbum Park <kese111@gmail.com>