mcount: Add 'libmcount/mcount-dynamic.c'
when use dynamic trace feature, 'uftrace' use 'libmcount-dynamic.so'
instead of 'libmcount.so'. 'mcount-dynamic.c' is implementation of
'libmcount-dynamic.so'. it hook the 'pre' and 'post' function from
'mcount.c' to do some work that needs for dynamic trace like parse
environment variable saved as file.

'Mcount-dynamic', which is responsible for Dynamic Trace, performs
almost all the operations required for Trace by using existing
functions implemented in 'mcount'. So, to minimize the effect on
'mcount', create the hooks and implemented it to use the required
for dynamic trace.

Modified Makefile to support build with 'libmcount/mcount-dynamic.c'.

Signed-off-by: Hanbum Park <kese111@gmail.com>