uftrace: check freepace of shared memory before run
When the shared memory space becomes full, running uftrace to trace
something will cause the "Bus error". This is rare case and is not
problem of uftrace.

because this case is rare, it will take time to find the actual problem
of an error. So, added a minimized logic that tells the user if the
shared memory is full.

Fixed: #371

Signed-off-by: Hanbum Park <kese111@gmail.com>